Move around tests for readability
